+++ composer.json
@@ -42,7 +42,8 @@
         "doctrine/doctrine-bundle": "^2.13",
         "doctrine/doctrine-migrations-bundle": "^3.3",
         "doctrine/orm": "^3.2",
-        "symfonycasts/reset-password-bundle": "^1.22"
+        "symfonycasts/reset-password-bundle": "^1.22",
+        "symfony/ux-turbo": "^2.20"
     },
     "require-dev": {
         "symfony/phpunit-bridge": "^7.1",
