+++ tailwindcss/preset.js
@@ -1,0 +1,11 @@
+const colors = require('tailwindcss/colors');
+
+module.exports = {
+    theme: {
+        extend: {
+            colors: {
+                primary: colors.blue,
+            }
+        },
+    },
+}
