+++ templates/components/DonationForm.html.twig
@@ -1,55 +1,3 @@
-<div {{ attributes }}>
-    {{ form_start(form, {'attr': {'novalidate': 'novalidate'}}) }}
-        <div class="flex flex-col">
-            {{ form_errors(form) }}
-            <div class="mb-3">
-                {# {{ form_row(form.type) }} #}
-                {{ form_label(form.amount, null , {'label_attr': {'class': 'text-xl text-center'}}) }}
-                {# {{ form_widget(form.amount) }} #}
-                {{ form_widget(form.amount, {'attr': {'class': 'p-4 text-2xl text-center'}}) }}
-                {{ form_errors(form.amount) }}
-                {{ form_help(form.amount) }}
-            </div>
-            {{ form_row(form.email) }}
-            {# {{ form_row(form.chosenAmount) }} #}
-            {% if form.taxReturn is defined %}
-                {{ form_row(form.taxReturn) }}
-                {% if form.taxReturn is same as true %}
-                    {{ form_row(form.givenName) }}
-                    {{ form_row(form.familyName) }}
-                    {{ form_row(form.nationalIdCode) }}
-                {% endif %}
-            {% endif %}
-            <div class="order-last">
-                <h1 class="font-medium text-xl text-center mb-3">Vali makseviis</h1>
-                {{ form_row(form.bankCountry) }}
-                <div class="mb-3">
-                    {% if form.gateway is defined %}
-                        {{ form_errors(form.gateway) }}
-                        <ul class="grid w-full gap-3 grid-cols-3">
-                            {% for choice in form.gateway.children %}
-                                <li>
-                                    <input
-                                        type="radio"
-                                        id="donation_gateway_{{ loop.index0 }}"
-                                        name="{{ field_name(form.gateway) }}"
-                                        value="{{ choice.vars.value }}"
-                                        placeholder="{{ choice.vars.label }}"
-                                        class="hidden peer"
-                                        {% if choice.vars.checked %} checked="checked"{% endif %}
-                                        required>
-                                    <label for="donation_gateway_{{ loop.index0 }}" class="aspect-square inline-flex items-center justify-between w-full p-3 text-gray-500 bg-white border border-gray-200 rounded-lg cursor-pointer dark:hover:text-gray-300 dark:border-gray-700 dark:peer-checked:text-blue-500 peer-checked:border-blue-600 peer-checked:text-blue-600 hover:text-gray-600 hover:bg-gray-100 dark:text-gray-400 dark:bg-gray-800 dark:hover:bg-gray-700">                           
-                                        <div class="block">
-                                            <img class="size-full" src="{{ asset(choice.vars.image) }}" alt="{{ choice.vars.label }}"/>
-                                        </div>
-                                    </label>
-                                </li>
-                            {% endfor %}
-                        </ul>
-                    {% endif %}
-                </div>
-                {{ form_row(form.submit, { 'label': 'Anneta', 'attr': {'class': 'w-full px-6 py-3.5 text-base font-medium text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:outline-none focus:ring-red-300 rounded-lg text-center dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-800'} }) }}
-            </class>
-        </div>
-    {{ form_end(form) }}
-</div>
+<turbo-frame id="donation-form">
+    {{ component('DonationFormStep' ~ step, {form: form, donationData: donationData, action: action}) }}
+</turbo-frame>
