+++ templates/components/DonationFormStep3.html.twig
@@ -1,0 +1,38 @@
+{% extends '@Donation/components/donation_form_layout.html.twig' %}
+
+{% set step = 3 %}
+
+{% block title %}Makse{% endblock %}
+
+{% block form_fields %}
+        <div class="flex flex-col">
+            {{ form_errors(form) }}
+            {{ form_row(form.bankCountry) }}
+            <div class="mb-3">
+                {% if form.gateway is defined %}
+                    {{ form_errors(form.gateway) }}
+                    <ul class="grid w-full gap-3 grid-cols-3">
+                        {% for choice in form.gateway.children %}
+                            <li>
+                                <input
+                                    type="radio"
+                                    id="donation_gateway_{{ loop.index0 }}"
+                                    name="{{ field_name(form.gateway) }}"
+                                    value="{{ choice.vars.value }}"
+                                    placeholder="{{ choice.vars.label }}"
+                                    class="hidden peer"
+                                    {% if choice.vars.checked %} checked="checked"{% endif %}
+                                    required>
+                                <label for="donation_gateway_{{ loop.index0 }}" class="aspect-square inline-flex items-center justify-between w-full p-3 text-gray-500 bg-white border border-gray-200 rounded-lg cursor-pointer dark:hover:text-gray-300 dark:border-gray-700 dark:peer-checked:text-blue-500 peer-checked:border-blue-600 peer-checked:text-blue-600 hover:text-gray-600 hover:bg-gray-100 dark:text-gray-400 dark:bg-gray-800 dark:hover:bg-gray-700">                           
+                                    <div class="block">
+                                        <img class="size-full" src="{{ asset(choice.vars.image) }}" alt="{{ choice.vars.label }}"/>
+                                    </div>
+                                </label>
+                            </li>
+                        {% endfor %}
+                    </ul>
+                {% endif %}
+            </div>
+            <button type="submit" class="w-full px-6 py-3.5 text-base font-medium text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:outline-none focus:ring-primary-300 rounded-lg text-center dark:bg-primary-600 dark:hover:bg-primary-700 dark:focus:ring-primary-800" data-turbo="false">Anneta</button>
+        </div>
+{% endblock %}
