+++ templates/components/donation_form_layout.html.twig
@@ -1,0 +1,39 @@
+<div class="container" {{ attributes }}>
+    {{ form_start(form, {'attr': {'novalidate': 'novalidate'}}) }}
+
+    {# This div is required to to wrap everything below into enclosing form tags properly #}
+    <div class="flex flex-col">
+
+        {# Header #}
+        <div class="flex flex-row">
+            <div class="size-14">
+                {% if step > 1 %}
+                <a class="inline-block p-4 rounded-lg hover:bg-gray-100 dark:bg-gray-800 dark:hover:bg-gray-700" href="/{{ step - 1}}">
+                    <svg class="w-6 h-6 text-gray-800 dark:text-white" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path d="M9.4 233.4c-12.5 12.5-12.5 32.8 0 45.3l160 160c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L109.2 288 416 288c17.7 0 32-14.3 32-32s-14.3-32-32-32l-306.7 0L214.6 118.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0l-160 160z"/></svg>
+                </a>
+                {% endif %}
+            </div>
+            <div class="grow flex flex-col">
+                <h1 class="text-xl text-center">
+                    {% block title %}{% endblock %}
+                </h1>
+                <div class="flex justify-center h-7">
+                    <span class="flex self-center w-3 h-3 m-1 {% if step >= 2 %}bg-primary-700{% else %}bg-gray-200{% endif %} rounded-full"></span>
+                    <span class="flex self-center w-3 h-3 m-1 {% if step >= 3 %}bg-primary-700{% else %}bg-gray-200{% endif %} rounded-full"></span>
+                    <span class="flex self-center w-3 h-3 m-1 bg-gray-200 rounded-full"></span>
+                </div>
+            </div>
+            <div class="size-14">
+                {% if step < 3 %}
+                <button type="submit" type="button" class="inline-block p-4 rounded-lg hover:bg-gray-100 dark:bg-gray-800 dark:hover:bg-gray-700">
+                    <svg class="w-6 h-6 text-gray-800 dark:text-white" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><!--!Font Awesome Free 6.6.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free Copyright 2024 Fonticons, Inc.--><path d="M438.6 278.6c12.5-12.5 12.5-32.8 0-45.3l-160-160c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3L338.8 224 32 224c-17.7 0-32 14.3-32 32s14.3 32 32 32l306.7 0L233.4 393.4c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0l160-160z"/></svg>
+                </button>
+                {% endif %}
+            </div>
+        </div>
+
+        {# Form fields #}
+        {% block form_fields %}{% endblock %}
+    </div>
+    {{ form_end(form) }}
+</div>
