+++ templates/thankyou.html.twig
@@ -1,0 +1,5 @@
+{% extends '@Donation/base.html.twig' %}
+
+{% block body %}
+    Thank you for your donation
+{% endblock %}
